Add NCCL/RCCL stubs and wrappers.
RCCL and NCCL have almost the same API (except for cudaStream_t vs hipStream_t) and use the same function names. The two are unbundled by prefixing the RCCL stub with `rccl` instead of `nccl`. The `rccl_stub.{h,cc}.inc` files are manually edited to achieve this.

The platform specific wrappers add very little value, but it's consistent with the other libraries and the code is mechanical and cheap to maintain.

PiperOrigin-RevId: 387673903