[TfrtTpuRunner] Add TPU memory manager to manage the TPU memory.
TpuMemoryManager is responsible for managing the TPU on-chip memory. Conceptually,
this is similar to the virtual memory manager of a typical OS which is
responsible for managing the host RAM and paging between RAM and disk.

In this CL, we introduce a very simple TPU memory manager that never evicts any
program and constant data loaded to TPU. We will add support for more
sophisticated memory manager that swaps the memory between host RAM and on-chip
memory in order to support multi-tenancy and larger ML models.

PiperOrigin-RevId: 390477819