Add `tfrt_gpu.ccl.unique_id/create` ops and ccl tests.
Discovered that NCCL insists on running on the primary context. I intend to add support for non-primary streams in NCCL upstream.

In the meantime (it might take a while), add checks and document the requirements. For the tests, add a `tfrt_gpu.context.primary` op.

Move the CclGroupStart/End() calls to GpuCclHandle::execute() to prevent captured buffer references to go away before CclGroupEnd() is called (which does the actual work).

PiperOrigin-RevId: 393097988